Move SkunkController class to the Domain layer
This is per Prof. Level's recommendation (see message below):

"Larman's Controller is part of the domain layer.  See Larman ch 27
which states this.

This controller is different than that within the Web MVC architecture,
where there it is part of the UI.

The Skunk controller is tricky to test, so do the best you can.  Other
tests look good."